NF: factorize onKeyDown
